Move useCartEntryToReplace to its own module
