+++ apps/store/src/components/ProductPage/PriceIntentContext.tsx
@@ -1,29 +1,21 @@
 'use client'
 import { useApolloClient } from '@apollo/client'
 import { datadogLogs } from '@datadog/browser-logs'
-import type {
-  PropsWithChildren} from 'react';
-import {
-  createContext,
-  useCallback,
-  useContext,
-  useEffect,
-  useState,
-} from 'react'
+import type { PropsWithChildren } from 'react'
+import { createContext, useCallback, useContext, useEffect, useState } from 'react'
 import { useProductData } from '@/components/ProductData/ProductDataProvider'
 import { useProductPageContext } from '@/components/ProductPage/ProductPageContext'
+import { useCartEntryToReplace } from '@/components/ProductPage/useCartEntryToReplace'
 import type {
   PriceIntentFragment,
   PriceIntentQueryResult,
-  ProductOfferFragment} from '@/services/graphql/generated';
-import {
-  usePriceIntentQuery,
+  ProductOfferFragment,
 } from '@/services/graphql/generated'
+import { usePriceIntentQuery } from '@/services/graphql/generated'
 import { priceIntentServiceInitClientSide } from '@/services/priceIntent/PriceIntentService'
 import type { ShopSession } from '@/services/shopSession/ShopSession.types'
 import { useShopSession } from '@/services/shopSession/ShopSessionContext'
 import { getOffersByPrice } from '@/utils/getOffersByPrice'
-import { useCartEntryToReplace } from './ProductPage'
 import { usePreloadedPriceIntentId } from './PurchaseForm/usePreloadedPriceIntentId'
 import { useSelectedOffer } from './PurchaseForm/useSelectedOffer'
