+++ apps/store/src/components/ProductPage/PurchaseForm/OfferPresenter.tsx
@@ -3,20 +3,17 @@ import { datadogRum } from '@datadog/browser-rum'
 import styled from '@emotion/styled'
 import { useInView } from 'framer-motion'
 import { useTranslation } from 'next-i18next'
-import type { RefObject} from 'react';
+import type { RefObject } from 'react'
 import { useEffect, useMemo, useRef, useState } from 'react'
 import { Button, Space, Text, theme } from 'ui'
 import { CancellationForm } from '@/components/Cancellation/CancellationForm'
 import { ScrollPast } from '@/components/ProductPage/ScrollPast/ScrollPast'
 import { ScrollToTopButton } from '@/components/ProductPage/ScrollToButton/ScrollToButton'
+import { useCartEntryToReplace } from '@/components/ProductPage/useCartEntryToReplace'
 import { SpaceFlex } from '@/components/SpaceFlex/SpaceFlex'
 import { BankSigneringEvent } from '@/services/bankSignering'
-import type {
-  ProductOfferFragment,
-  RedeemedCampaignFragment} from '@/services/graphql/generated';
-import {
-  ExternalInsuranceCancellationOption
-} from '@/services/graphql/generated'
+import type { ProductOfferFragment, RedeemedCampaignFragment } from '@/services/graphql/generated'
+import { ExternalInsuranceCancellationOption } from '@/services/graphql/generated'
 import type { PriceIntent } from '@/services/priceIntent/priceIntent.types'
 import type { ShopSession } from '@/services/shopSession/ShopSession.types'
 import { useTracking } from '@/services/Tracking/useTracking'
@@ -26,7 +23,6 @@ import { PageLink } from '@/utils/PageLink'
 import { useAddToCart } from '@/utils/useAddToCart'
 import { useGetDiscountExplanation } from '@/utils/useDiscountExplanation'
 import { useFormatter } from '@/utils/useFormatter'
-import { useCartEntryToReplace } from '../ProductPage'
 import { ComparisonTableModal } from './ComparisonTableModal'
 import { DeductibleSelector } from './DeductibleSelector'
 import { DiscountTooltip } from './DiscountTooltip/DiscountTooltip'
